requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
psycopg2-binary==2.9.9
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
